requirements.txt
streamlit==1.28.1
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.1.78
lime==0.2.0.1
grad-cam==1.4.8
scikit-learn==1.3.0
tqdm==4.66.1
